vertex_ai/speech/speech_to_text/transcribe/requirements.txt
google-genai
google-cloud-secret-manager
google-adk
google-api-python-client
google-auth-oauthlib
requests-oauthlib
google-auth
gunicorn
Flask
Flask-Cors
google-cloud-resource-manager
mistune
Quart
scipy
google-cloud-texttospeech
Flask-Sockets
fastapi
uvicorn[standard]
websockets